app/src/main/java/com/example/calculatorapp/domains/calculator/infrastructure/specifications/CalculatorSpecifications.java
package com.example.calculatorapp.domains.calculator.infrastructure.specifications;

import static com.example.calculatorapp.constants.StringUtilitiesConstants.*;
import static com.example.calculatorapp.constants.domains.CalculatorConstants.*;

import java.util.*;

public class CalculatorSpecifications {
    private CalculatorSpecifications() {
    }

    public static boolean isCalculationResultARationalNumber(double evaluatedCalculationExpression) {
        return evaluatedCalculationExpression != (int) evaluatedCalculationExpression;
    }

    public static boolean isCalculationExpressionEmpty(String calculationExpression) {
        return Objects.equals(calculationExpression, EMPTY_STRING);
    }

    public static boolean isCalculationExpressionNotValidExpressionExceptionMessage(String calculationExpression) {
        return Objects.equals(calculationExpression, NOT_VALID_EXPRESSION_EXCEPTION_MESSAGE);
    }
}